Resolve 4 conflicts:
- translations/base.json: merged both sides' keys (deduped the shared "Rate
  card" key; kept the 4 admin filter labels)
- MainNavMenuSections: combined the empty-group guard (main's catalogTabs +
  the admin adminTabs)
- Filters/graphql/utils.ts: combined the id-label chip switch cases
  (adminOrganizations + main's rateCardProduct/rateCardProductFilter)
- graphql.tsx: regenerated against the backend (admin schema + main's latest)

Migrate the admin pages/components off the removed `react-router-dom` package
onto `react-router` (main's v7 rename): admin pages, AdminGuard, and the
Navigate import in core/router.

+++ src/pages/CreditNotesPage.tsx
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client'
 import { useMemo, useState } from 'react'
-import { useSearchParams } from 'react-router-dom'
+import { useSearchParams } from 'react-router'
 
 import CreditNotesTable from '~/components/creditNote/CreditNotesTable'
 import { usePageSearchParam } from '~/components/designSystem/Pagination'
@@ -34,8 +34,8 @@ import { useOrganizationInfos } from '~/hooks/useOrganizationInfos'
 
 gql`
   query getCreditNotesList(
-    $amountFrom: Int
-    $amountTo: Int
+    $amountFrom: BigInt
+    $amountTo: BigInt
     $creditStatus: [CreditNoteCreditStatusEnum!]
     $currency: CurrencyEnum
     $customerExternalId: String
